+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Fixes
 
-- Remove x/y coordinates from UI tap breadcrumbs, as they are a potential security risk, for example leaking input on custom PIN code views (#8534)
+- Remove x/y coordinates from UI breadcrumbs, as they are a potential security risk, for example leaking input on custom PIN code views (#8534)
 
 ## 9.23.0